test: move tokens test from `PublishV1Test.swift` to `PublisherTest.swift`
